JBuilder should probably bet set to use the javac compiler, not its own internal borland java compiler (in JBuilder, see Project->Project Properties->Build).
Fixed a bug that manifested itself only with JBuilder (with NetBeans it wasn't there).  If getColumnName(i) in the table model returns "" (empty string) for the first N columns, the other columns names in the header get "smushed" also.  i.e. the column header height becomes only a few pixels and the names for the other columns are not really visible.  To fix this, just have getColumnName() return " " instead of "" for those columns which do not have a name.