Bug 834 : change the behaviour of the CLI with respect to config files.
If you don't specify one, all the options must be supplied on the command line.

If you do specify one, options can be overridden on the command line, but they will not be written into the config file.

If you don't specify one and use -G to start the gui, the default configuration file will be written and used as usual: it will be modified by other options you pass in or change in the gui.

If you do specify one and use -G to start the gui, the specified config file will NOT be modified.

This allows a consistent reuse of the customized config files, if so desired.